Make links under species map consistent with map filter
e.g. add fq=occurrence_status:present if only presence records are displayed on the species map.
Spatial portal link is slightly different due to quirk in its use of the first parameter as legend entry and disregarding of subsequent parameters if these are used in a second or later call for the same user.